🔄 Many new changes abound
I was going to scrap all of the css and start from scratch, but I created a good baseline of reusable code and it felt wasteful to throw it out, so I am working with it. Added some placeholder images on about and work. Still figuring out layout, but its starting to look well enough.